requirements.txt
pytorch-lightning==1.9
pytorch_metric_learning
torch_geometric
jsonargparse
